[MODIFIED & ADDED]: Folder structure of server side modified and new middleware added
