+++ .gitignore
@@ -1,15 +1,5 @@
-.env
 node_modules
+.env
 dist
 coverage
 ```
-
-
-
-
-
-
-
-
-
-
