+++ package.json
@@ -8,15 +8,16 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
-  "author": "",
+  "author": "Rafi, Jabed, Sushanto, Imran, Abdur Rahman, Shoyon",
   "license": "ISC",
-  "description": "",
+  "description": "AI-powered course management system backend built with Node.js, Express.js, and MongoDB, following the MVC architecture for scalable and secure API development.",
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.15.0",
-    "mongoose": "^8.12.1"
+    "mongoose": "^8.12.1",
+    "morgan": "^1.10.0"
   }
 }
