+++ src/controllers/userController.js
@@ -1,4 +1,4 @@
-const User = require("../models/userModels");
+const User = require("../models/User");
 const { sendResponse } = require("../utils");
 
 const getUsers = async (req, res, next) => {
@@ -10,11 +10,20 @@ const getUsers = async (req, res, next) => {
   }
 };
 
-
 const createUser = async (req, res, next) => {
   try {
     const userData = req.body;
-    const result = await User.insertOne(userData);
+
+    const isUser = await User.findOne({ email: userData.email });
+    console.log({ isUser });
+    
+    let result;
+    if (!isUser) {
+      result = await User.create(userData);
+    } else {
+      result = isUser;
+    }
+
     sendResponse(res, 201, true, "Successfully users created", result);
   } catch (error) {
     next(error); // pass the error to the global handler
@@ -25,5 +34,3 @@ module.exports = {
   getUsers,
   createUser
 }
-
-
